Added magnifying glass 'searchable' icon on db A-Z list for searchable dbs. Can be easily disabled by variable in local xsl/includes.xsl, example included in demo/xsl/includes.xsl
Added search box on db detail page for searchable db. Can be easily disabled
by variable in demo/xsl/includes.xsl, example included. Search box shows up if
db is searchable by logged in or on campus user, or is searchable by someone and
user is not logged in or on campus. Searchable dbs we know user doesn't have access to,
a message is displayed.