Change architecture for keeping track of saved records from a result set. Previously this was kept track of in a special cookie. Now it is kept track of in session. Previously when looking at a result set, records were marked as already saved by javascript run on page load. Now they are marked as already saved in HTML generation (although javascript can still modify for ajax functionality, of course).
This has some functional improvements for browsers without javascript. The change also, pleasingly, results in a reduction of lines of code in Xerxes, I think, it's simpler.

But the main reason to do it was to set the stage for adding label control to the results page for saved records, doing things this way will make adding an ajax (or non ajax, even) tag labler directly on the search results page much easier.

This commit also fixes a bug with MetasearchSearch involving spell correction and a typo.