Added label input box functionality on the search results page for saved records.
Technical what was changed:

Labelling input form moved to it's own XSL template in includes.xsl. 

search results xsl will generate label inputs. This is good for non-js. To do this, lookup of labels of saved records and putting them in XML was added to commands (which were already looking up that a record was saved from last change).  These label inputs will work fine without javascript, going to the 'your labels were updated' page, with a link back to results. 

For AJAX, in save.js the routine that takes care of saving records was changed to add a label area on save, and remove it on delete. To do this, at the bottom of the page a hidden div is provided with a 'template' label input. The js, on save, will make a copy of this template div, then fill out all the fields properly for the record in question. To do that, it needs the Xerxes_Record id for the saved record---the label edit action was updated to provide a json response with this information. 

Also on the search results page in a hidden div is the label list/menu, of all labels the given user has. This is provided to make the tag.js javascript happy, because it expects such a div. But more importantly, to be used for initializing an autocomplete list for tag entry, to be done.  To support this, the tag listing command was added to search results action.